GVT-2973: Refactor PublicationUtils to multiple files
