+++ infra/src/main/kotlin/fi/fta/geoviite/infra/publication/PublicationController.kt
@@ -21,6 +21,7 @@ import fi.fta.geoviite.infra.localization.LocalizationService
 import fi.fta.geoviite.infra.localization.localizationParams
 import fi.fta.geoviite.infra.util.Page
 import fi.fta.geoviite.infra.util.SortOrder
+import fi.fta.geoviite.infra.util.getCsvResponseEntity
 import fi.fta.geoviite.infra.util.toResponse
 import java.time.Duration
 import java.time.Instant
